Fix sequential reading

+++ Position.h
@@ -2,7 +2,7 @@
 #define POSITION_H
 #include <stdio.h>
 typedef struct position *Position;
-Position POSITIONinit(int x, int y);
+Position POSITIONinit(double x, double y);
 void POSITIONfree(Position p);
 double POSITIONcompute_euclidean_distance(Position p1, Position p2);
 int POSITIONcmp(Position p1, Position p2);
